Generalised the confirm_delete() Javascript helper function and added wrappers for deleting Reports and Blogposts
--HG--
branch : blogpost support

+++ Projects/TDoR/src/views/posts/delete.php
@@ -3,18 +3,24 @@
      * Delete the current Blogpost.
      *
      */
+    require_once('util/blog_utils.php');
 
-    //require_once('models/blog_utils.php');
 
-
-    //if (is_editor_user() )
+    if (is_admin_user() )
     {
         $db             = new db_credentials();
         $posts_table    = new Posts($db);
 
         if ($posts_table->delete($post) )
         {
-            echo "Blogpost $post->title deleted";
+            echo '<h2>Blogposts</h2>';
+            echo '<div><img src="/images/tdor_candle_jars.jpg" /></div>';
+
+            echo '<p>&nbsp;</p>';
+            echo get_top_level_menu_html();         // Top level menu
+            echo '<p>&nbsp;</p>';
+
+            echo "Blogpost \"$post->title\" deleted";
         }
     }
